added field deletion route

+++ backend/src/main/java/com/playvu/backend/service/SubFieldService.java
@@ -10,12 +10,15 @@ import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.playvu.backend.entity.SubField;
 import com.playvu.backend.entity.Users;
 import com.playvu.backend.repository.FieldRepository;
 import com.playvu.backend.repository.FieldScheduleRepository;
+import com.playvu.backend.repository.GameRepository;
 import com.playvu.backend.repository.SubFieldRepository;
 
 import jakarta.servlet.http.HttpServletRequest;
@@ -32,18 +35,21 @@ public class SubFieldService {
     private UserService userService;
 
     @Autowired
+    private GameRepository gameRepository;
+
+    @Autowired
     private FieldScheduleRepository fieldScheduleRepository;
 
-    public Integer addSubField(HttpServletRequest request, Integer field_id, String name) throws URISyntaxException, IOException, InterruptedException{
+    public Integer addSubField(HttpServletRequest request, Integer fieldId, String name) throws URISyntaxException, IOException, InterruptedException{
 
         Users user = userService.getUserFromJwt();
 
-        if(fieldRepository.findById(field_id).get().getOwnerId() != user.getUserId()){
-            return null;
+        if(fieldRepository.findById(fieldId).get().getOwnerId() != user.getUserId()){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "User does not control field: " + fieldId);
         }
 
         SubField new_subfield = new SubField();
-        new_subfield.setMasterFieldId(field_id);
+        new_subfield.setMasterFieldId(fieldId);
 
         new_subfield.setName(name);
 
@@ -58,8 +64,13 @@ public class SubFieldService {
 
         Integer masterFieldId = subFieldRepository.findBySubFieldId(subFieldId).getMasterFieldId();
         if(fieldRepository.findById(masterFieldId).get().getOwnerId() != user.getUserId()){
-            return;
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "User does not control subfield: " + subFieldId);
         }
+        System.out.println(gameRepository.isPendingGame(subFieldId));
+        if(gameRepository.isPendingGame(subFieldId)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Pending game in subfield: " + subFieldId);
+        }
+        
 
         subFieldRepository.deleteById(subFieldId);
     }
